autoload all classes in forms directory lazy-loads forms models when used normalize require paths require 'src/forms' to autoload forms classes instead of requiring individual forms directly.
