+++ dashboard/lib/api/v1/pd/workshop_score_summarizer.rb
@@ -1,4 +1,4 @@
-require_dependency pegasus_dir('forms/pd_workshop_survey')
+require_dependency pegasus_dir('src/forms')
 
 module Api::V1::Pd::WorkshopScoreSummarizer
   FACILITATOR_EFFECTIVENESS_QUESTIONS = [
