+++ pegasus/forms/class_submission.rb
@@ -1,5 +1,3 @@
-require_relative './form'
-
 class ClassSubmission < Form
   def self.normalize(data)
     result = {}
