+++ pegasus/forms/company_profile.rb
@@ -1,5 +1,3 @@
-require_relative './form'
-
 class CompanyProfile < Form
   def self.normalize(data)
     {}.tap do |result|
