+++ pegasus/forms/pd_program_registration.rb
@@ -1,5 +1,3 @@
-require_relative './form'
-
 class PdProgramRegistration < Form
   def self.normalize(data)
     result = {}
