+++ pegasus/forms/pd_workshop_survey_counselor_admin.rb
@@ -1,5 +1,3 @@
-require_relative './form'
-
 class PdWorkshopSurveyCounselorAdmin < Form
   def self.normalize(data)
     result = {}
