+++ pegasus/forms/volunteer_contact_2015.rb
@@ -1,5 +1,3 @@
-require_relative './form'
-
 class VolunteerContact2015 < Form
   def self.normalize(data)
     result = {}
