+++ pegasus/forms/volunteer_engineer_submission.rb
@@ -1,5 +1,3 @@
-require_relative './form'
-
 class VolunteerEngineerSubmission < Form
   def self.normalize(data)
     result = {}
