+++ pegasus/src/forms.rb
@@ -2,6 +2,10 @@ require 'cdo/poste'
 require 'cdo/regexp'
 require src_dir 'database'
 require lib_dir 'forms/pegasus_form_errors'
+require lib_dir 'forms/pegasus_form_validation'
+require 'active_support/core_ext/string/inflections'
 
-# Load forms
-Dir.glob(pegasus_dir('forms/*.rb')).each {|path| require path}
+# Autoload all classes in forms directory.
+Dir.glob(pegasus_dir('forms/*.rb')).each do |path|
+  autoload File.basename(path, '.rb').camelize, path
+end
